lab3/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pika
python-dotenv